requirements.txt
fastapi[all]
pyjwt
passlib
pymongo
python-multipart